Remove nested controller for forms, initialize params
The value of the select for the Region/LGA/Ward dropdown is bound to
a params model. If that model doesn't exist on the controller scope,
it will be created on the wrong scope when the select is first changed,
causing the value to be ignored. We need to initialize params on the
controller scope immediately when the controller loads s.t. the model
is found and set in the right scope.